SquishyNavbar: If power saver is on, don't allow feedback
When power saver is on, the squeeze vibration is turned off. This is done
by Google to save battery life across the system. Keeping the navbar animation
on doesn't make sense if we're trying to save power. This commit fixes that.

- Also, make sure we're not calling the navbar animation while the screen
is turned off. Although we don't see it, we're technically still calling on
it on.